Adjust levels of Vale rules
Change the error levels of Vale style rules to pave the way for
requiring the Vale linter.

If a style rule has the `raw` scope, it is not currently possible to
disable the rule within a document using HTML/MDX comments. Further, a
rule with the `error` level causes Vale to exit with a nonzero error
code if a page violates the rule. As a result, this change makes the
following adjustments to the error levels of Vale style rules:

- `error`: styles where we want to ensure consistency/correctness, and
  where it is possible to disable the style using comments.
- `warning`: style rules that an author might need to ignore. This
  includes style rules with the `raw` scope, which cannot be disabled
  using comments.

For some style rules, change the scope from `raw` to another scope that
allows a docs author to disable the rule within a document using
HTML/MDX-style comments.